crawler/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
amqp==5.1.1
apache-libcloud==3.7.0
async-timeout==4.0.2
attrs==23.1.0
billiard==4.1.0
blinker==1.6.2
celery==5.3.1
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.5
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
discord.py==2.3.1
Flask==2.3.2
frozenlist==1.4.0
greenlet==2.0.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.3.1
MarkupSafe==2.1.3
multidict==6.0.4
mysql-connector-python==8.0.33
prompt-toolkit==3.0.39
protobuf==3.20.3
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0
redis==4.6.0
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.18
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.3
vine==5.0.0
wcwidth==0.2.6
Werkzeug==2.3.6
yarl==1.9.2
